requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
pandas==2.2.2
openpyxl==3.1.5
